+++ src/AnalyticsManager.php
@@ -1114,7 +1114,7 @@ class AnalyticsManager
      */
     public function version(): string
     {
-        return '2.16.0';
+        return '2.17.0';
     }
 
     /**